pyproject.toml
[project]
name = "mapfisher"
version = "0.1.0"
description = "CLI fishing game that also serves as a weather app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-staticmaps >= 0.4.0",
    "pillow >= 12.0.0",
    "img2unicode >= 0.1a11"
]
license = "GPL-3.0-only"
license-files = ["LICENSE"]

[build-system]
requires = ["uv_build >= 0.7.19, <0.9.0"]
build-backend = "uv_build"

[project.scripts]
mapfish = "mapfisher.main:entry"
mapfisher = "mapfisher.main:entry"

[tool.setuptools.packages.find]
where = ["src"]